Fix the Hanagumi Hall ring vent overflow
The way the cap was being handled was kind of messed up, and was trying to subtract your existing Super Ring count.
But because the datatype doesn't support negative numbers, it rolled under to approx UINT16_MAX!
Fix the datatypes and add the extra over-range check to prevent this from happening again.