Battle UFO: fix spawn ordering logic
If the mapthings went in reverse order, so mapthing #1 had
a higher ID than mapthing #2, the spawn ordering logic
would get stuck because it wanted mapthing #1 to have
a lower ID than mapthing #2.

The root of the bug is that the previously used spawners
ID was not totally ignored. Now it is.