MT_FLINGRING: Nullify fuse on touch
If the fuse caused the flingring to be deleted, it would hog a spot on player->pickuprings until the player respawned.

This is not a perfect solution, but it should resolve the problem in practice.

Needs testing to confirm this doesn't introduce TA desync, but in the worst case I would wager it affects a single digit number of runs at worst.