Turn off water ripples when reducevfx is on
A number of people have said they have sensory issues caused by the
ripple effects from water and heat. This turns off those effects
when reducevfx is on.

It may be a better idea for this to be a separate option, but it's
a starting point we can build upon later.