Use integer arithmetic for pwrlv avg calculation
This fixes an oversight with pwrlv average calculation that causes
the total to overflow with enough players or high enough pwrlv.
Hopefully this might fix that bug where pwrlv is shown as negative
on the server select menu.

Maintainer note: This is still imprecise but it fixes the overflow
without potentially disrupting game code.