Disable camera dampening when looking backwards.
When traveling quickly, camera dampening normally causes the camera to get
further away from the player, but it screws up camera angles when looking
backwards.

This change causes camera dampening to happen instantly when looking behind.
That way, players can expect the camera to always be in a consistent location
when they tap the "look behind" button.