Don't free stringargs during deep copy
spawnsectors is memcpy'd over just before the calls to copy_stringarg,
resulting in copy_stringarg Z_Free'ing the exact string it is about to
copy. This is a silent memory error, as the malloc block appears to
still be in a virtual-allocated page after the call to Z_Free.

deepcopy is not able to properly free the target array's tree of objects
if the target already exists, so that will need future work.